#1 - Add Processors
Added more tests

+++ MetadataExtractor/Metadata.cs
@@ -5,7 +5,7 @@
 
     public class Metadata
     {
-        public uint? FocalLengthIn35MmFormat { get; set; }
+        public ushort? FocalLengthIn35MmFormat { get; set; }
         public DateTime? CaptureTime { get; set; }
         public string CameraManufacturer { get; set; }
         public string CameraModel { get; set; }
@@ -28,7 +28,7 @@
         public string EditingSoftware { get; set; }
         public DateTime? ModifiedTime { get; set; }
         public DateTime? CreationTime { get; set; }
-        public uint? Iso { get; set; }
+        public ushort? Iso { get; set; }
         public decimal? ExposureCompensation { get; set; }
         public MeteringModeEnum? MeteringMode { get; set; }
         public ExposureModeEnum? ExposureMode { get; set; }
